+++ src/routes/student.route.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 
-const studentController = require('./controller');
+const studentController = require('../controllers/student.controller');
 
 router.get('/', async (req, res) => {
   try {
